# backend_db/backend/openapi_extracted.yaml
openapi: 3.1.0
info:
  title: SelfDB
  description: Full Self-Hosted BaaS Built for AI Agents.
  version: 0.0.5
paths:
...
  /tables/:
    post:
      tags:
      - tables
      summary: Create Table
      description: Create a new table. Requires authentication.
      operationId: create_table_tables__post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
    get:
      tags:
      - tables
      summary: List Tables
      description: 'List tables with optional search and sorting.


        - **search**: Filter by name or description (case-insensitive, printable ASCII
        only)

        - **sort_by**: Field to sort by (created_at, updated_at, name)

        - **sort_order**: Sort direction (asc or desc, default: desc)


        Returns:

        - For unauthenticated users: Only public tables

        - For authenticated users: Public tables + tables owned by the user'
      operationId: read_tables_tables__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          maximum: 2147483647
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
            pattern: ^[\x20-\x7E]*$
          - type: 'null'
          description: Search term for filtering by name or description
          title: Search
        description: Search term for filtering by name or description
      - name: sort_by
        in: query
        required: false
        schema:
          enum:
          - created_at
          - updated_at
          - name
          type: string
          description: Field to sort by
          default: created_at
          title: Sort By
        description: Field to sort by
      - name: sort_order
        in: query
        required: false
        schema:
          enum:
          - asc
          - desc
          type: string
          description: Sort order (ascending or descending)
          default: desc
          title: Sort Order
        description: Sort order (ascending or descending)
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableRead'
                title: Response Read Tables Tables  Get
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    TableCreate:
      properties:
        name:
          type: string
          maxLength: 63
          minLength: 1
          pattern: ^[a-z][a-z0-9_]*$
          title: Name
          description: Table name (lowercase letters, numbers, underscores, must start
            with letter)
        table_schema:
          additionalProperties: true
          type: object
          title: Table Schema
          description: JSON schema definition for table structure
        public:
          type: boolean
          title: Public
          default: false
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      additionalProperties: false
      type: object
      required:
      - name
      - table_schema
      title: TableCreate
      examples:
      - description: Product catalog table
        metadata:
          version: '1.0'
        name: products
        public: true
        table_schema:
          properties:
            name:
              type: string
            price:
              type: number
          type: object
...
    endpoints__schema__ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TableRead:
      properties:
        name:
          type: string
          maxLength: 63
          minLength: 1
          pattern: ^[a-z][a-z0-9_]*$
          title: Name
        table_schema:
          additionalProperties: true
          type: object
          title: Table Schema
          description: JSON schema definition for table structure
        public:
          type: boolean
          title: Public
          default: false
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
        id:
          type: string
          format: uuid
          title: Id
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        row_count:
          type: integer
          title: Row Count
          default: 0
        realtime_enabled:
          type: boolean
          title: Realtime Enabled
          default: false
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - name
      - table_schema
      - id
      - owner_id
      title: TableRead
      examples:
      - description: Product catalog table
        name: products
        public: true
        table_schema:
          properties:
            name:
              type: string
          type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /users/token